Removed Surveyee concept
