+++ src/main/java/ca/group20/sysc4806project/repository/SurveyRepo.java
@@ -1,7 +1,7 @@
 package ca.group20.sysc4806project.repository;
 
-import ca.group20.sysc4806project.model.Surveyor;
 import ca.group20.sysc4806project.model.Survey;
+import ca.group20.sysc4806project.model.Surveyor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SurveyRepo extends JpaRepository<Survey, Long> {
