+++ src/main/java/ca/group20/sysc4806project/service/SurveyService.java
@@ -1,7 +1,6 @@
 package ca.group20.sysc4806project.service;
 
 import ca.group20.sysc4806project.model.Survey;
-import ca.group20.sysc4806project.model.Surveyor;
 
 public interface SurveyService {
     Survey saveSurvey(Survey survey);
